adk-demo-engine/requirements.txt
google-adk[eval]
Flask
pytest
mock
